+++ flux-common/src/main/java/com/danielgmyers/flux/threads/ThreadUtils.java
@@ -14,7 +14,7 @@
  *   limitations under the License.
  */
 
-package com.danielgmyers.flux.clients.swf.util;
+package com.danielgmyers.flux.threads;
 
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.atomic.AtomicInteger;
