+++ flux-swf/src/main/java/com/danielgmyers/flux/clients/swf/FluxCapacitorImpl.java
@@ -43,13 +43,13 @@ import com.danielgmyers.flux.FluxCapacitor;
 import com.danielgmyers.flux.RemoteWorkflowExecutor;
 import com.danielgmyers.flux.WorkflowStatusChecker;
 import com.danielgmyers.flux.clients.swf.poller.ActivityTaskPoller;
-import com.danielgmyers.flux.clients.swf.poller.BlockOnSubmissionThreadPoolExecutor;
 import com.danielgmyers.flux.clients.swf.poller.DecisionTaskPoller;
 import com.danielgmyers.flux.clients.swf.util.RetryUtils;
-import com.danielgmyers.flux.clients.swf.util.ThreadUtils;
 import com.danielgmyers.flux.poller.TaskNaming;
 import com.danielgmyers.flux.step.StepAttributes;
 import com.danielgmyers.flux.step.WorkflowStep;
+import com.danielgmyers.flux.threads.BlockOnSubmissionThreadPoolExecutor;
+import com.danielgmyers.flux.threads.ThreadUtils;
 import com.danielgmyers.flux.wf.Periodic;
 import com.danielgmyers.flux.wf.Workflow;
 import com.danielgmyers.flux.wf.graph.WorkflowGraphNode;
