+++ flux-swf/src/main/java/com/danielgmyers/flux/clients/swf/poller/ActivityTaskPoller.java
@@ -24,6 +24,7 @@ import com.danielgmyers.flux.clients.swf.util.RetryUtils;
 import com.danielgmyers.flux.poller.TaskNaming;
 import com.danielgmyers.flux.step.StepResult;
 import com.danielgmyers.flux.step.WorkflowStep;
+import com.danielgmyers.flux.threads.BlockOnSubmissionThreadPoolExecutor;
 import com.danielgmyers.flux.wf.Workflow;
 import com.danielgmyers.metrics.MetricRecorder;
 import com.danielgmyers.metrics.MetricRecorderFactory;
