+++ flux-swf/src/main/java/com/danielgmyers/flux/clients/swf/poller/DecisionTaskPoller.java
@@ -35,7 +35,6 @@ import com.danielgmyers.flux.clients.swf.FluxCapacitorImpl;
 import com.danielgmyers.flux.clients.swf.IdentifierValidation;
 import com.danielgmyers.flux.clients.swf.poller.timers.TimerData;
 import com.danielgmyers.flux.clients.swf.util.RetryUtils;
-import com.danielgmyers.flux.clients.swf.util.ThreadUtils;
 import com.danielgmyers.flux.poller.TaskNaming;
 import com.danielgmyers.flux.signals.BaseSignalData;
 import com.danielgmyers.flux.signals.DelayRetrySignalData;
@@ -47,6 +46,8 @@ import com.danielgmyers.flux.step.StepAttributes;
 import com.danielgmyers.flux.step.StepResult;
 import com.danielgmyers.flux.step.WorkflowStep;
 import com.danielgmyers.flux.step.WorkflowStepUtil;
+import com.danielgmyers.flux.threads.BlockOnSubmissionThreadPoolExecutor;
+import com.danielgmyers.flux.threads.ThreadUtils;
 import com.danielgmyers.flux.wf.Periodic;
 import com.danielgmyers.flux.wf.Workflow;
 import com.danielgmyers.flux.wf.graph.WorkflowGraph;
