+++ flux-swf/src/test/java/com/danielgmyers/flux/clients/swf/poller/ActivityTaskPollerTest.java
@@ -34,6 +34,7 @@ import com.danielgmyers.flux.step.StepApply;
 import com.danielgmyers.flux.step.StepAttributes;
 import com.danielgmyers.flux.step.StepResult;
 import com.danielgmyers.flux.step.WorkflowStep;
+import com.danielgmyers.flux.threads.BlockOnSubmissionThreadPoolExecutor;
 import com.danielgmyers.flux.wf.Workflow;
 import com.danielgmyers.flux.wf.graph.WorkflowGraph;
 import com.danielgmyers.flux.wf.graph.WorkflowGraphBuilder;
