+++ flux-swf/src/test/java/com/danielgmyers/flux/clients/swf/poller/WorkflowStateTest.java
@@ -36,7 +36,6 @@ import com.danielgmyers.flux.clients.swf.poller.testwf.TestWorkflow;
 import com.danielgmyers.flux.clients.swf.poller.testwf.TestWorkflowWithFailureTransition;
 import com.danielgmyers.flux.clients.swf.poller.testwf.TestWorkflowWithPartitionedStep;
 import com.danielgmyers.flux.clients.swf.poller.timers.TimerData;
-import com.danielgmyers.flux.clients.swf.util.ManualClock;
 import com.danielgmyers.flux.poller.TaskNaming;
 import com.danielgmyers.flux.signals.DelayRetrySignalData;
 import com.danielgmyers.flux.signals.ForceResultSignalData;
@@ -44,6 +43,7 @@ import com.danielgmyers.flux.signals.SignalType;
 import com.danielgmyers.flux.step.PartitionIdGeneratorResult;
 import com.danielgmyers.flux.step.StepAttributes;
 import com.danielgmyers.flux.step.StepResult;
+import com.danielgmyers.flux.testutil.ManualClock;
 import com.danielgmyers.flux.wf.Workflow;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
